Update sample app
