+++ sample/ACPAnalyticsSample/android/app/BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.acpcoresample",
+    package = "com.acpanalyticssample",
 )
 
 android_resource(
     name = "res",
-    package = "com.acpcoresample",
+    package = "com.acpanalyticssample",
     res = "src/main/res",
 )
