+++ sample/ACPAnalyticsSample/android/app/src/main/java/com/acpanalyticssample/MainApplication.java
@@ -1,26 +1,15 @@
-/*
-Copyright 2019 Adobe. All rights reserved.
-This file is licensed to you under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License. You may obtain a copy
-of the License at http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software distributed under
-the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
-OF ANY KIND, either express or implied. See the License for the specific language
-governing permissions and limitations under the License.
-*/
-package com.acpcoresample;
+package com.acpanalyticssample;
 
 import android.app.Application;
 
 import com.facebook.react.ReactApplication;
-import com.adobe.marketing.mobile.reactnative.RCTACPCorePackage;
 import com.adobe.marketing.mobile.reactnative.analytics.RCTACPAnalyticsPackage;
+import com.adobe.marketing.mobile.reactnative.RCTACPCorePackage;
 import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
 import com.facebook.react.shell.MainReactPackage;
 import com.facebook.soloader.SoLoader;
-import com.adobe.marketing.mobile.reactnative.RCTACPCoreModule;
+import com.adobe.marketing.mobile.MobileCore; // import MobileCore
 
 import java.util.Arrays;
 import java.util.List;
@@ -37,8 +26,8 @@ public class MainApplication extends Application implements ReactApplication {
     protected List<ReactPackage> getPackages() {
       return Arrays.<ReactPackage>asList(
           new MainReactPackage(),
-            new RCTACPCorePackage(),
-            new RCTACPAnalyticsPackage()
+            new RCTACPAnalyticsPackage(),
+            new RCTACPCorePackage()
       );
     }
 
@@ -57,6 +46,6 @@ public class MainApplication extends Application implements ReactApplication {
   public void onCreate() {
     super.onCreate();
     SoLoader.init(this, /* native exopackage */ false);
-    RCTACPCoreModule.setApplication(this);
+    MobileCore.setApplication(this); // add this line 
   }
 }
