+++ sample/ACPAnalyticsSample/android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">ACPCoreSample</string>
+    <string name="app_name">ACPAnalyticsSample</string>
 </resources>
