+++ sample/ACPAnalyticsSample/android/settings.gradle
@@ -1,7 +1,7 @@
-rootProject.name = 'ACPCoreSample'
-include ':@adobe_react-native-acpcore'
-project(':@adobe_react-native-acpcore').projectDir = new File(rootProject.projectDir, '../node_modules/@adobe/react-native-acpcore/android')
+rootProject.name = 'ACPAnalyticsSample'
 include ':@adobe_react-native-acpanalytics'
 project(':@adobe_react-native-acpanalytics').projectDir = new File(rootProject.projectDir, '../node_modules/@adobe/react-native-acpanalytics/android')
+include ':@adobe_react-native-acpcore'
+project(':@adobe_react-native-acpcore').projectDir = new File(rootProject.projectDir, '../node_modules/@adobe/react-native-acpcore/android')
 
 include ':app'
