+++ sample/ACPAnalyticsSample/app.json
@@ -1,4 +1,4 @@
 {
-  "name": "ACPCoreSample",
-  "displayName": "ACPCoreSample"
+  "name": "ACPAnalyticsSample",
+  "displayName": "ACPAnalyticsSample"
 }
