+++ sample/ACPAnalyticsSample/babel.config.js
@@ -1,3 +1,3 @@
 module.exports = {
-  presets: ["module:metro-react-native-babel-preset"]
-}
+  presets: ['module:metro-react-native-babel-preset'],
+};
