+++ sample/ACPAnalyticsSample/ios/ACPAnalyticsSample-tvOSTests/Info.plist
