+++ sample/ACPAnalyticsSample/ios/ACPAnalyticsSample/Images.xcassets/AppIcon.appiconset/Contents.json
