+++ sample/ACPAnalyticsSample/ios/ACPAnalyticsSample/Images.xcassets/Contents.json
