+++ sample/ACPAnalyticsSample/ios/ACPAnalyticsSampleTests/ACPAnalyticsSampleTests.m
@@ -1,14 +1,9 @@
-/*
-Copyright 2019 Adobe. All rights reserved.
-This file is licensed to you under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License. You may obtain a copy
-of the License at http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software distributed under
-the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
-OF ANY KIND, either express or implied. See the License for the specific language
-governing permissions and limitations under the License.
-*/
+/**
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
 
 #import <UIKit/UIKit.h>
 #import <XCTest/XCTest.h>
@@ -19,11 +14,11 @@ governing permissions and limitations under the License.
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React Native!"
 
-@interface ACPCoreSampleTests : XCTestCase
+@interface ACPAnalyticsSampleTests : XCTestCase
 
 @end
 
-@implementation ACPCoreSampleTests
+@implementation ACPAnalyticsSampleTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
