+++ sample/ACPAnalyticsSample/ios/ACPAnalyticsSampleTests/Info.plist
